+++ src/js/PlotBoilerplate.d.ts
@@ -64,7 +64,7 @@ import { BezierPath } from "./BezierPath";
 import { Grid } from "./Grid";
 import { Polygon } from "./Polygon";
 import { Vertex } from "./Vertex";
-import { Bounds, Config, Drawable, DrawConfig, IHooks, PBParams, XYCoords, XYDimension } from "./interfaces";
+import { IBounds, Config, Drawable, DrawConfig, IHooks, PBParams, XYCoords, XYDimension } from "./interfaces";
 /**
  * A wrapper class for draggable items (mostly vertices).
  * @private
@@ -451,7 +451,7 @@ export declare class PlotBoilerplate {
      * @memberof PlotBoilerplate
      * @return {Bounds} The current viewport.
      **/
-    viewport(): Bounds;
+    viewport(): IBounds;
     /**
      * Trigger the saveFile.hook.
      *